Create functioning live server
Browser sync automatically runs, but only on a server whose port is 3100.
The server is set to the port 3100.
To run browser-sync, you have to go to port 3000 when you open the page.

Modify the startAll command to be npm-run-all '--parallel'.
'Parallel' is important because npm commands run sychronously.
So Webpack would wait for Nodemon to finish, and it would throw an error because Nodemon would never finish.

Also added /www to .gitignore